backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyMySQL==1.1.0
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pandas==2.1.4
python-dotenv==1.0.0
authlib==1.3.0
requests==2.31.0
isodate==0.6.1
firebase-admin==6.2.0
